make functions work better together
changed add_year call in prototype_totals read in the
dataframe instead of the list of prototypes. add_year was changed to
not initialize the column, removes messages of writing to a copy
of the dataframe. rx_commission_decommission resets the index at
the end so the Time column isn't the index if there are no reactors
decommissioned