fix bug in add_receivr_prototype
Bug was in the merging of the resource and transactions tables.
Originally it was removing transations of resources with duplicate
ResourceId, removing some of the transactions from the merged
DataFrame. This fix is for the assignment of the wrong value from
the Quantity column of Resources to the merged df.