update pu equivalence calculations, add keff comparison
Update the pu equivalence calculation to get the correct
relationship between the quivalence factor and the actinide
weight fraction. Add a section to run transport with changed
fuel composition (UOX to MOX or U/TRU to HALEU) for
the SMR and SFR to compare the keff values. Change to
reflective boundary conditions for SFR lattices to do the
single assembly transport